Allow admin to add new invoice + payment in dashboard
Did not work before, because the form in the table
couldn't be submitted, because the resulting HTML was
invalid. Instead an event listener to the button is now
attached.
Also we now not only create a new invoice, but also directly
add a new payment with the same amount as the previously created
invoice