Fix overflow on team page
This fixes an overflow on the team page. For some reason
(I suspect the text backing the icon) the delete icon on
postings causes an overflow, which also leads to the navbar
being wider than the screen.

Hiding this overflow and making the icon a little wider to
display it fully solves this problem